HW8 -Rework meal.jsp to datatable     -Implement add/delete action meal by using Bootstrap and ajax     -Rework delete button to use with onClick action

+++ src/main/webapp/resources/js/topjava.common.js
@@ -3,11 +3,6 @@ var context, form;
 function makeEditable(ctx) {
     context = ctx;
     form = $('#detailsForm');
-    $(".delete").click(function () {
-        if (confirm('Are you sure?')) {
-            deleteRow($(this).attr("id"));
-        }
-    });
 
     $(document).ajaxError(function (event, jqXHR, options, jsExc) {
         failNoty(jqXHR);
@@ -23,6 +18,9 @@ function add() {
 }
 
 function deleteRow(id) {
+    if (!confirm('Are you sure?'))
+        return
+
     $.ajax({
         url: context.ajaxUrl + id,
         type: "DELETE"
